requirements.txt
python-telegram-bot[job-queue]==20.*
python-dotenv
aiohttp
paramiko
Pillow
google-generativeai
